Rename cache configuration and packages
org.seedstack.seed.cache -> org.seedstack.cache
