+++ src/it/java/org/seedstack/jcache/SampleExpiryPolicyFactory.java
@@ -5,7 +5,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.seedstack.seed.cache;
+package org.seedstack.jcache;
 
 import javax.cache.configuration.Factory;
 import javax.cache.expiry.Duration;
