+++ src/main/java/org/seedstack/jcache/internal/CacheConcern.java
@@ -5,7 +5,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.seedstack.seed.cache.spi;
+package org.seedstack.jcache.internal;
 
 
 import io.nuun.kernel.spi.Concern;
@@ -14,11 +14,11 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 
 /**
- * Nuun concern for ordering cache operations.
+ * Nuun concern for ordering jcache operations.
  *
  * @author adrien.lauer@mpsa.com
  */
 @Retention(RetentionPolicy.RUNTIME)
-@Concern(name="seed-cache-concern", priority= Concern.Priority.HIGH)
+@Concern(name="jcache-concern", priority= Concern.Priority.HIGH)
 public @interface CacheConcern {
 }
