+++ src/main/resources/META-INF/services/io.nuun.kernel.api.Plugin
@@ -1,1 +1,1 @@
-org.seedstack.seed.cache.internal.CachePlugin
+org.seedstack.jcache.internal.CachePlugin
